+++ src/include/nfs_core.h
@@ -599,7 +599,7 @@ extern writeverf3 NFS3_write_verifier; /* NFS V3 write verifier */
 extern nfs_worker_data_t *workers_data;
 extern char config_path[MAXPATHLEN];
 extern char pidfile_path[MAXPATHLEN] ;
-extern ushort g_nodeid;
+extern int g_nodeid;
 
 typedef enum process_status
 {
