+++ src/include/sal_functions.h
@@ -793,7 +793,7 @@ nfs4_create_clid_name(nfs_client_record_t * cl_recp,
 void nfs4_add_clid(nfs_client_id_t *);
 void nfs4_rm_clid(char *);
 void nfs4_chk_clid(nfs_client_id_t *);
-void nfs4_load_recov_clids(ushort nodeid);
+void nfs4_load_recov_clids(int nodeid);
 void nfs4_clean_old_recov_dir();
 void nfs4_create_recov_dir();
